Rename `testFiles` to more general `config`.
This change will allow various configuration information about the
program to be stored in a single place.
